+++ js/affiliation.js
@@ -467,6 +467,22 @@ var Affiliation = {
       },
     },
 
+    'omegav': { // Subaffiliation of Omega
+      name: 'Omega Verksted',
+      key: 'omegav',
+      web: 'http://omegav.no/',
+      logo: './org/omegav/logo.png',
+      icon: './org/omegav/icon.png',
+      symbol: './org/omegav/symbol.png',
+      placeholder: './org/omegav/placeholder.png',
+      palette: 'grey',
+      news: {
+        type: 'feed',
+        feed: 'http://omegav.no/newsrss',
+        imageScraping: {},
+      },
+    },
+
     'online': {
       name: 'Online',
       key: 'online',
@@ -630,69 +646,61 @@ var Affiliation = {
         imageScraping: {newsSelector:'div#main', domainUrl:'smorekoppen.no/'},
       },
     },
-    // // TODO: Put Industrivinduet into Smørekoppen as an alternative news source specific to them
-    // 'industrivinduet': {
-    //   name: 'Industrivinduet',
-    //   key: 'industrivinduet',
-    //   web: 'http://www.smorekoppen.no/?subsite=industrivinduet',
-    //   // feed not available, use getNews instead
-    //   logo: './org/industrivinduet/logo.png',
-    //   icon: './org/industrivinduet/icon.png',
-    //   symbol: './org/industrivinduet/symbol.png',
-    //   placeholder: './org/industrivinduet/placeholder.png',
-    //   palette: 'blue',
-    //   getImage: function(link, callback) {
-    //     Images.get(this, link, callback, {newsSelector:'div#main', domainUrl:'smorekoppen.no/'});
-    //   },
-    //   getNews: function(posts, callback) {
-    //     var self = this;
-    //     Ajaxer.getCleanHtml({
-    //       url: self.web,
-    //       success: function(html) {
-    //         var aff = Affiliation.org['industrivinduet'];
-    //         var count = 0;
 
-    //         // Add each item from news tags
-    //         if ($(html).find('li[id^="article-"]').length != 0) {
-    //           $(html).find('li[id^="article-"]').each( function() {
-    //             if (count < posts.length) {
-    //               var post = {};
-                  
-    //               // The popular fields
-    //               post.title = $(this).find('h2 a').text();
-    //               post.link = $(this).find('h2 a').attr('href');
-    //               post.description = $(this).eq(0).find('p.subline').text().trim().replace(/\s+/g, ' ');
-    //               post.author = $(this).next().find('p.subline').text().trim();
-    //               post.image = $(this).find('pic').attr('src');
-
-    //               // Author fixing
-    //               post.author = post.author.match(/[a-zæøå\-'_]+ [a-zæøå\-'_]+/i);
-    //               if (post.author !== null)
-    //                 post.author = post.author[0];
-    //               else
-    //                 post.author = 'A/F Smørekoppen';
-
-    //               // Image fixing
-    //               if (typeof post.image === 'undefined')
-    //                 post.image = aff.placeholder;
-    //               else
-    //                 post.image = aff.web + post.image;
-
-    //               posts[count++] = post;
-    //             }
-    //           });
-    //         }
-    //         else {
-    //           console.error('No articles found at', aff.web);
-    //         }
-    //         callback(posts);
-    //       },
-    //       error: function(e) {
-    //         console.error('Could not fetch '+aff.name+' website');
-    //       },
-    //     });
-    //   },
-    // },
+    'industrivinduet': { // Subaffiliation of A/F Smørekoppen
+      name: 'Industrivinduet',
+      key: 'industrivinduet',
+      web: 'http://www.smorekoppen.no/?subsite=industrivinduet',
+      logo: './org/industrivinduet/logo.png',
+      icon: './org/industrivinduet/icon.png',
+      symbol: './org/industrivinduet/symbol.png',
+      placeholder: './org/industrivinduet/placeholder.png',
+      palette: 'blue',
+      news: {
+        type: 'website',
+        scrape: function(html, limit, callback) {
+          var aff = Affiliation.org['industrivinduet'];
+          var count = 0;
+          var posts = [];
+          // Add each item from news tags
+          if ($(html).find('li[id^="article-"]').length != 0) {
+            $(html).find('li[id^="article-"]').each( function() {
+              if (count < limit) {
+                var post = {};
+                // The popular fields
+                post.title = $(this).find('h2 a').text();
+                post.link = $(this).find('h2 a').attr('href');
+                post.description = $(this).eq(0).find('p.subline').text().trim().replace(/\s+/g, ' ');
+                post.author = $(this).next().find('p.subline').text().trim();
+                post.image = $(this).find('pic').attr('src');
+                // Author fixing
+                post.author = post.author.match(/[a-zæøå\-'_]+ [a-zæøå\-'_]+/i);
+                if (post.author !== null) {
+                  post.author = post.author[0];
+                }
+                else {
+                  post.author = 'Industrivinduet';
+                }
+                // Image fixing
+                if (typeof post.image === 'undefined') {
+                  post.image = aff.placeholder;
+                }
+                else {
+                  // Add domain URL because it's missing.
+                  // Add "whatever=.jpg" to help the image past Images.control
+                  post.image = 'http://smorekoppen.no/' + post.image + '&isimage=.jpg';
+                }
+                posts[count++] = post;
+              }
+            });
+          }
+          else {
+            console.error('No articles found at', aff.web);
+          }
+          callback(posts);
+        },
+      },
+    },
 
     'solan': {
       name: 'Solan',
@@ -804,6 +812,22 @@ var Affiliation = {
     //
     // Linjeforeninger Dragvoll
     //
+
+    'caf': {
+      name: 'CAF',
+      key: 'caf',
+      web: 'https://cafidrett.wordpress.com/',
+      logo: './org/caf/logo.png',
+      icon: './org/caf/icon.png',
+      symbol: './org/caf/symbol.png',
+      placeholder: './org/caf/placeholder.png',
+      palette: 'blue',
+      news: {
+        type: 'feed',
+        feed: 'https://cafidrett.wordpress.com/feed/',
+        imageScraping: {},
+      },
+    },
 
     'communitas': {
       name: 'Communitas',